SQL data-source - refactored and separated out
The multi-targeting within a single file had got too much for me
to cognitively process, so separated out the files for .NET Framework
and .NET Core.

Adds support for non-Umbraco databases on v10/v11/v12.
