+++ src/Umbraco.Community.Contentment/DataEditors/DataList/DataSources/SqlDataListSource.NET472.cs
@@ -1,0 +1,120 @@
+﻿/* Copyright © 2019 Lee Kelleher.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+/* NOTE: This code file is ONLY for Umbraco v8, .NET Framework 4.7.2.
+ * For the .NET Core 5.0 version, please see file `SqlDataListSource.NET5_0.cs`. */
+
+#if NET472
+using System.Collections.Generic;
+using System.Configuration;
+using System.Data.Common;
+using System.Data.SqlClient;
+using System.Data.SqlServerCe;
+using System.Linq;
+using Microsoft.AspNetCore.Hosting;
+using Umbraco.Core;
+using Umbraco.Core.IO;
+using UmbConstants = Umbraco.Core.Constants;
+
+namespace Umbraco.Community.Contentment.DataEditors
+{
+    public sealed partial class SqlDataListSource : IDataListSource
+    {
+        private readonly string _codeEditorMode;
+        private readonly IEnumerable<DataListItem> _connectionStrings;
+        private readonly IIOHelper _ioHelper;
+
+        public SqlDataListSource(
+            IWebHostEnvironment webHostEnvironment,
+            IIOHelper ioHelper)
+        {
+            // NOTE: Umbraco doesn't ship with SqlServer mode, so we check if its been added manually, otherwise defautls to Razor.
+            _codeEditorMode = webHostEnvironment.WebPathExists("~/umbraco/lib/ace-builds/src-min-noconflict/mode-sqlserver.js") == true
+                ? "sqlserver"
+                : "razor";
+
+            _connectionStrings = ConfigurationManager.ConnectionStrings
+                .Cast<ConnectionStringSettings>()
+                .Select(x => new DataListItem
+                {
+                    Name = x.Name,
+                    Value = x.Name
+                });
+
+            _ioHelper = ioHelper;
+        }
+
+        public IEnumerable<DataListItem> GetItems(Dictionary<string, object> config)
+        {
+            var items = new List<DataListItem>();
+
+            var query = config.GetValueAs("query", string.Empty);
+            var connectionStringName = config.GetValueAs("connectionString", string.Empty);
+
+            if (string.IsNullOrWhiteSpace(query) == true || string.IsNullOrWhiteSpace(connectionStringName) == true)
+            {
+                return items;
+            }
+
+            var settings = ConfigurationManager.ConnectionStrings[connectionStringName];
+            if (settings == null)
+            {
+                return items;
+            }
+
+            // NOTE: SQLCE uses a different connection/command. I'm trying to keep this as generic as possible, without resorting to using NPoco. [LK]
+            if (settings.ProviderName.InvariantEquals(UmbConstants.DatabaseProviders.SqlCe) == true)
+            {
+                items.AddRange(GetSqlItems<SqlCeConnection, SqlCeCommand>(query, settings.ConnectionString));
+            }
+            else
+            {
+                items.AddRange(GetSqlItems<SqlConnection, SqlCommand>(query, settings.ConnectionString));
+            }
+
+            return items;
+        }
+
+        private IEnumerable<DataListItem> GetSqlItems<TConnection, TCommand>(string query, string connectionString)
+            where TConnection : DbConnection, new()
+            where TCommand : DbCommand, new()
+        {
+            using (var connection = new TConnection() { ConnectionString = connectionString })
+            using (var command = new TCommand() { Connection = connection, CommandText = query })
+            {
+                connection.Open();
+                using (var reader = command.ExecuteReader())
+                {
+                    while (reader.Read() == true)
+                    {
+                        if (reader.FieldCount > 0)
+                        {
+                            var item = new DataListItem
+                            {
+                                Name = reader[0].TryConvertTo<string>().Result
+                            };
+
+                            item.Value = reader.FieldCount > 1
+                                ? reader[1].TryConvertTo<string>().Result
+                                : item.Name;
+
+                            if (reader.FieldCount > 2)
+                                item.Description = reader[2].TryConvertTo<string>().Result;
+
+                            if (reader.FieldCount > 3)
+                                item.Icon = reader[3].TryConvertTo<string>().Result;
+
+                            if (reader.FieldCount > 4)
+                                item.Disabled = reader[4].ToString().TryConvertTo<bool>().Result;
+
+                            yield return item;
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+#endif
